+++ src/Enum/Join.php
@@ -7,4 +7,5 @@ enum Join: string
     case INNER = 'INNER JOIN';
     case LEFT = 'LEFT JOIN';
     case RIGHT = 'RIGHT JOIN';
+    case FULL = 'FULL JOIN';
 }