+++ src/Query/Debugger.php
@@ -243,7 +243,8 @@ class Debugger
         $keywords = [
             'SELECT', 'FROM', 'WHERE', 'ORDER', 'GROUP', 'BY', 'HAVING', 'DISTINCT', 'EXCLUDE',
             'LIMIT', 'OFFSET', 'JOIN', 'ON', 'AS', 'AND', 'OR', 'DESC', 'LIKE', 'XOR',
-            'ASC', 'IN', 'IS', 'NOT', 'NULL', 'LEFT', 'INNER'
+            'ASC', 'IN', 'IS', 'NOT', 'NULL', 'LEFT', 'INNER', 'RIGHT', 'FULL', 'OUTER',
+            'CASE', 'WHEN', 'THEN', 'ELSE', 'END'
         ];
 
         // Function: Uppercase letters, numbers and underscores, at least 2 characters, cannot start/end with underscore