+++ src/Traits/Joinable.php
@@ -52,6 +52,12 @@ trait Joinable
         return $this;
     }
 
+    public function fullJoin(Query $query, string $alias): Query
+    {
+        $this->addJoin($query, Enum\Join::FULL, $alias);
+        return $this;
+    }
+
     public function on(string $leftKey, Enum\Operator $operator, string $rightKey): Query
     {
         $joinKey = array_key_last($this->joins);